# packages/wos-api/openapi.yaml
openapi: 3.1.0
info:
  title: Woobe Outcome Server HTTP API
  version: 0.1.0-dev
  description: |
    HTTP contract implemented through Wave 07. The API exposes the same
    Application services used by embedded consumers. Namespace and Outcome
    scope are explicit in every domain path. Issues and Blockers are independent
    aggregates; compound coordination commands remain explicit and atomic.
servers:
  - url: http://127.0.0.1:8080
...
paths:
...
  /api/v1/namespaces/{namespace_id}/outcomes:
    post:
      tags: [Outcomes]
      operationId: createOutcome
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOutcomeRequest"
      responses:
        "201":
          description: Outcome created or replayed idempotently.
          headers:
            ETag:
              schema: {type: string}
            Location:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutcomeMutationResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/namespaces/{namespace_id}/outcomes/{outcome_id}:
    get:
      tags: [Outcomes]
      operationId: getOutcome
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/CorrelationID"
      responses:
        "200":
          description: Persisted Outcome representation.
          headers:
            ETag:
              schema: {type: string}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutcomeReadResult"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Outcomes]
      operationId: updateOutcome
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOutcomeRequest"
      responses:
        "200":
          $ref: "#/components/responses/OutcomeMutation"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /api/v1/namespaces/{namespace_id}/outcomes/{outcome_id}/state:
    get:
      tags: [Outcomes]
      operationId: getOutcomeState
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/CorrelationID"
      responses:
        "200":
          description: Coherent Outcome state projection including Wave 07 blocking state.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutcomeState"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/namespaces/{namespace_id}/outcomes/{outcome_id}/actions/activate:
    post:
      tags: [Outcomes]
      operationId: activateOutcome
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        $ref: "#/components/requestBodies/VersionBody"
      responses:
        "200":
          $ref: "#/components/responses/OutcomeMutation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
  /api/v1/namespaces/{namespace_id}/outcomes/{outcome_id}/actions/achieve:
    post:
      tags: [Outcomes]
      operationId: achieveOutcome
      parameters:
        - $ref: "#/components/parameters/NamespaceID"
        - $ref: "#/components/parameters/OutcomeID"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        $ref: "#/components/requestBodies/ReasonBody"
      responses:
        "200":
          $ref: "#/components/responses/OutcomeMutation"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
...
components:
  parameters:
    NamespaceID:
      name: namespace_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/ID"}
    OutcomeID:
      name: outcome_id
      in: path
      required: true
      schema: {$ref: "#/components/schemas/ID"}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: |
        Strong ETag of the persisted aggregate, for example
        "work_item:0199e100-0000-7000-8000-000000000001:v5".
        Either If-Match or expected_version in the body is required for
        versioned mutations. If both are present they must agree.
      schema: {type: string}
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema: {type: string}
  requestBodies:
    VersionBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/VersionRequest"
    ReasonBody:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ReasonRequest"
...
  responses:
    OutcomeMutation:
      description: Outcome mutation confirmed or replayed.
      headers:
        ETag:
          schema: {type: string}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OutcomeMutationResult"
    BadRequest:
      description: Invalid request, identifier, JSON or header format.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found in the explicit Namespace/Outcome scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Lifecycle, lease, criterion or idempotency conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    PreconditionFailed:
      description: Persisted version did not satisfy the supplied precondition.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    PreconditionRequired:
      description: A required If-Match or expected_version precondition is absent.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    ID:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
    Version:
      type: integer
      minimum: 1
    Priority:
      type: string
      enum: [critical, high, normal, low]
    OutcomeLifecycle:
      type: string
      enum: [draft, active, achieved, failed, abandoned]
    ObjectiveLifecycle:
      type: string
      enum: [planned, in_progress, achieved, cancelled]
    WorkItemLifecycle:
      type: string
      enum: [backlog, todo, in_progress, done, cancelled]
    RelationLifecycle:
      type: string
      enum: [active, removed]
    DependencyStrength:
      type: string
      enum: [hard, advisory]
    DependencySatisfaction:
      type: string
      enum: [target_completed]
    IssueSeverity:
      type: string
      enum: [critical, major, minor, informational]
    IssueLifecycle:
      type: string
      enum: [open, investigating, resolved, wont_fix, duplicate]
    BlockerLifecycle:
      type: string
      enum: [active, resolved, cancelled]
    BlockerPropagation:
      type: string
      enum: [direct, subtree]
...
    CreateOutcomeRequest:
      type: object
      additionalProperties: false
      required: [title, desired_state, priority]
      properties:
        title: {type: string}
        description: {type: string}
        desired_state: {type: string}
        priority: {$ref: "#/components/schemas/Priority"}
...
    ExternalCause:
      type: object
      additionalProperties: false
      required: [provider, description]
      properties:
        provider: {type: string}
        id: {type: string}
        uri: {type: string, format: uri}
        description: {type: string}
...
    UpdateOutcomeRequest:
      type: object
      additionalProperties: false
      properties:
        expected_version: {$ref: "#/components/schemas/Version"}
        title: {type: string}
        description: {type: string}
        desired_state: {type: string}
        priority: {$ref: "#/components/schemas/Priority"}
...
    VersionRequest:
      type: object
      additionalProperties: false
      properties:
        expected_version: {$ref: "#/components/schemas/Version"}
    ReasonRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        expected_version: {$ref: "#/components/schemas/Version"}
        reason: {type: string}
...
    ActorRef:
      type: object
      required: [kind, provider, id]
      properties:
        kind:
          type: string
          enum: [human, agent, service, automation, external_system]
        provider: {type: string}
        id: {type: string}
        display_name: {type: string}
...
    Outcome:
      type: object
      required: [id, namespace_id, version, title, desired_state, lifecycle, priority, criteria, created_at, updated_at]
      properties:
        id: {$ref: "#/components/schemas/ID"}
        namespace_id: {$ref: "#/components/schemas/ID"}
        version: {$ref: "#/components/schemas/Version"}
        title: {type: string}
        description: {type: string}
        desired_state: {type: string}
        lifecycle: {$ref: "#/components/schemas/OutcomeLifecycle"}
        priority: {$ref: "#/components/schemas/Priority"}
        criteria: {type: object}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Objective:
      type: object
      required: [id, scope, version, title, lifecycle, priority, required_for_outcome, criteria, created_at, updated_at]
      properties:
        id: {$ref: "#/components/schemas/ID"}
        scope: {type: object}
        version: {$ref: "#/components/schemas/Version"}
        title: {type: string}
        description: {type: string}
        lifecycle: {$ref: "#/components/schemas/ObjectiveLifecycle"}
        priority: {$ref: "#/components/schemas/Priority"}
        required_for_outcome: {type: boolean}
        criteria: {type: object}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    WorkItem:
      type: object
      required: [id, scope, version, title, lifecycle, priority, last_fencing_token, criteria, created_at, updated_at]
      properties:
        id: {$ref: "#/components/schemas/ID"}
        scope: {type: object}
        version: {$ref: "#/components/schemas/Version"}
        title: {type: string}
        description: {type: string}
        objective_id: {$ref: "#/components/schemas/ID"}
        lifecycle: {$ref: "#/components/schemas/WorkItemLifecycle"}
        priority: {$ref: "#/components/schemas/Priority"}
        not_before: {type: string, format: date-time}
        result_summary: {type: string}
        current_lease: {type: object}
        last_fencing_token: {type: integer, minimum: 0}
        criteria: {type: object}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Relation:
      type: object
      required: [id, scope, version, source_ref, relation_type, target_ref, lifecycle, created_at, updated_at]
      properties:
        id: {$ref: "#/components/schemas/ID"}
        scope: {type: object}
        version: {$ref: "#/components/schemas/Version"}
        source_ref: {type: object}
        relation_type:
          type: string
          enum: [depends_on, relates_to, produces, derived_from]
        target_ref: {type: object}
        strength: {$ref: "#/components/schemas/DependencyStrength"}
        satisfaction: {$ref: "#/components/schemas/DependencySatisfaction"}
        lifecycle: {$ref: "#/components/schemas/RelationLifecycle"}
        reason: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Issue:
      type: object
      required: [id, scope, version, title, severity, affected_refs, lifecycle, reported_by, created_at, updated_at]
      properties:
        id: {$ref: "#/components/schemas/ID"}
        scope: {type: object}
        version: {$ref: "#/components/schemas/Version"}
        title: {type: string}
        description: {type: string}
        severity: {$ref: "#/components/schemas/IssueSeverity"}
        affected_refs:
          type: array
          items: {type: object}
        lifecycle: {$ref: "#/components/schemas/IssueLifecycle"}
        reported_by: {$ref: "#/components/schemas/ActorRef"}
        resolution_summary: {type: string}
        duplicate_of_issue_id: {$ref: "#/components/schemas/ID"}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Blocker:
      type: object
      required: [id, scope, version, blocked_ref, lifecycle, propagation, created_at, updated_at]
      properties:
        id: {$ref: "#/components/schemas/ID"}
        scope: {type: object}
        version: {$ref: "#/components/schemas/Version"}
        blocked_ref: {type: object}
        cause_ref: {type: object}
        external_cause: {$ref: "#/components/schemas/ExternalCause"}
        description: {type: string}
        lifecycle: {$ref: "#/components/schemas/BlockerLifecycle"}
        propagation: {$ref: "#/components/schemas/BlockerPropagation"}
        resolved_at: {type: string, format: date-time}
        resolution_summary: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    AppliedBlocker:
      type: object
      required: [blocker, inherited]
      properties:
        blocker: {$ref: "#/components/schemas/Blocker"}
        inherited: {type: boolean}
        inherited_from: {type: object}
    BlockingState:
      type: object
      required: [ref, is_blocked, active_blockers]
      properties:
        ref: {type: object}
        is_blocked: {type: boolean}
        active_blockers:
          type: array
          items: {$ref: "#/components/schemas/AppliedBlocker"}
...
    WorkItemOperationalState:
      type: object
      required: [ref, lifecycle, display_state, lease_status, is_blocked, evaluated_at]
      properties:
        ref: {type: object}
        lifecycle: {$ref: "#/components/schemas/WorkItemLifecycle"}
        display_state:
          type: string
          enum: [backlog, waiting_scope, blocked, waiting_dependencies, scheduled, ready, in_progress, attention_needed, done, cancelled]
        lease_status:
          type: string
          enum: [none, active, expired]
        is_blocked: {type: boolean}
        readiness_reasons:
          type: array
          items:
            type: string
            enum:
              - lifecycle_not_ready
              - outcome_not_active
              - outcome_archived
              - objective_terminal
              - blocked
              - hard_dependency_unsatisfied
              - not_before
              - active_lease
              - lease_expired
        evaluated_at: {type: string, format: date-time}
...
    MutationMetadata:
      type: object
      required: [outcome_revision, command_id, idempotent_replay]
      properties:
        outcome_revision: {type: integer, minimum: 1}
        command_id: {$ref: "#/components/schemas/ID"}
        idempotent_replay: {type: boolean}
    OutcomeMutationResult:
      allOf:
        - $ref: "#/components/schemas/MutationMetadata"
        - type: object
          required: [value]
          properties:
            value: {$ref: "#/components/schemas/Outcome"}
...
    OutcomeReadResult:
      type: object
      required: [value, outcome_revision]
      properties:
        value: {$ref: "#/components/schemas/Outcome"}
        outcome_revision: {type: integer, minimum: 1}
...
    OutcomeState:
      type: object
      required: [outcome, objectives, work_items, work_item_operational_states, relations, issues, blockers, blocking_states, evaluated_at, outcome_revision]
      properties:
        outcome: {$ref: "#/components/schemas/Outcome"}
        objectives:
          type: array
          items: {$ref: "#/components/schemas/Objective"}
        work_items:
          type: array
          items: {$ref: "#/components/schemas/WorkItem"}
        work_item_operational_states:
          type: array
          items: {$ref: "#/components/schemas/WorkItemOperationalState"}
        relations:
          type: array
          items: {$ref: "#/components/schemas/Relation"}
        issues:
          type: array
          items: {$ref: "#/components/schemas/Issue"}
        blockers:
          type: array
          items: {$ref: "#/components/schemas/Blocker"}
        blocking_states:
          type: array
          items: {$ref: "#/components/schemas/BlockingState"}
        evaluated_at: {type: string, format: date-time}
        outcome_revision: {type: integer, minimum: 1}
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: {type: string}
            message: {type: string}
            retryable: {type: boolean}
            correlation_id: {type: string}